mysqli implementation - Successful Test
In this code there is an option to create a new table based on the
structure of another table.  Also there is a temporary varient.  This
varient is only active before it is closed… it is not permanently saved
to the disk.  This is useful for manupulating data before saving it to
a disk.