HTML Tag Tweaks & New CSS
Made minor adjustments to the tags contained within the printf
functions, to be certain that the printed tables are properly formatted
when checking a page's source code.

Also added a template of my CSS file as deployed in my CSIS-6 class of
2011.  Hopefully with this sheet style included in the project, it will
reduce the amount of formatting needed in the PHP code.