Procedural of index.php & include directory
Procedural of index.php has changed so that CSS is loaded and
displayed.  But if the login to the SQL database isn't successful then
the page loading stops to prevent any odd or confusing text to be
displayed in the browser.

Also the include directory contains some of the classes needed for this
project (user, supervisor), in additiion to some of the other classes
authored by me during class.